+++ Battleships.Core/Ship.cs
@@ -4,13 +4,13 @@ namespace Battleships.Core
 {
     public class Ship
     {
-        public byte NumberOfMasts { get; }
+        public int NumberOfMasts { get; }
         internal bool IsSunk => _hitMastsCount == NumberOfMasts;
         internal virtual void ReportHit() => _hitMastsCount++;
 
-        private byte _hitMastsCount;
+        private int _hitMastsCount;
 
-        public Ship(byte numberOfMasts)
+        public Ship(int numberOfMasts)
         {
             if (numberOfMasts == 0) throw new ArgumentException("Ship must have at least one mast.");
